backend/pyproject.toml
[tool.poetry]
name = "nfsp-donation-reader-be"
version = "0.1.0"
description = ""
authors = ["Luis Gustavo <luisgustavolf@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
pytesseract = "^0.3.13"
pillow = "^10.4.0"
starlette = "^0.38.2"
pydantic = "^2.8.2"
fastapi = {extras = ["standard"], version = "^0.112.2"}
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
